fix: resolve multiple regression integration test failures (#611)
This commit fixes several critical issues in the regression module:

- Fix DecisionTree weighted training to properly set IsLeaf and SplitValue
- Fix GradientBoosting to build trees sequentially (not parallel)
- Fix M5ModelTree serialization to include LinearModel
- Fix GAM serialization to properly handle base class data
- Fix RandomForest seed reproducibility in parallel execution
- Fix ARModel.Predict to use column as time series instead of rows
- Improve LogisticRegression gradient descent training

Test results improved from 97 passed/51 failed to 137 passed/11 failed.

+++ src/Regression/DecisionTreeAsyncRegressionBase.cs
@@ -261,7 +261,8 @@ public abstract class AsyncDecisionTreeRegressionBase<T> : IAsyncTreeBasedModel<
         // Serialize options
         writer.Write(Options.MaxDepth);
         writer.Write(Options.MinSamplesSplit);
-        writer.Write(double.IsNaN(Options.MaxFeatures) ? -1 : Options.MaxFeatures);
+        // Write MaxFeatures as int (-1 for NaN, cast to int otherwise)
+        writer.Write(double.IsNaN(Options.MaxFeatures) ? -1 : (int)Options.MaxFeatures);
         writer.Write(Options.Seed ?? -1);
         writer.Write((int)Options.SplitCriterion);
